reportGen/docWrite/pyproject.toml
[project]
name = "docwrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "strands-agents>=1.1.0",
    "strands-agents-tools>=0.2.2",
    "typing>=3.10.0.0",
]
